+++ src/BoundingRegion.cpp
@@ -38,9 +38,13 @@
 #include <CGAL/Delaunay_triangulation_2.h>
 
 BoundingRegion::BoundingRegion() = default;
+BoundingRegion::BoundingRegion(std::shared_ptr<Config::ReconRegion> reconRegion)
+: _reconSettings(reconRegion)
+{}
 BoundingRegion::~BoundingRegion() = default;
 
 //-- Operators to read bounded region if explicitly defined in config
+//   called through boost:apply_visitor
 void BoundingRegion::operator()(double radius) {
     geomutils::make_round_poly(global::nullPt, radius, _boundingRegion);
 }
@@ -49,35 +53,41 @@ void BoundingRegion::operator()(Polygon_2& poly) {
     _boundingRegion = poly;
 }
 
-void
+//-- Reconstruction (influ) regions related functions
+double
 BoundingRegion::calc_influ_region_bpg(const DT& dt, BuildingsPtr& buildings) {
-    assert(boost::get<bool>(Config::get().influRegionConfig));
     double influRegionRadius;
     //-- Find building where the point of interest lies in and define radius of interest with BPG
-    bool foundBuilding = false;
     for (auto& f : buildings) {
         if (geomutils::point_in_poly(global::nullPt, f->get_poly())) {
             f->calc_footprint_elevation_nni(dt);
             try {
+                f->set_reconstruction_rules(*this); // temporary add settings for this reconstruction
                 f->reconstruct();
+                f->remove_reconstruction_rules();
             } catch (std::exception& e) {
                 std::cerr << std::endl << "Error: " << e.what() << std::endl;
                 throw std::runtime_error("Impossible to automatically determine influence region");
             }
-            influRegionRadius = sqrt(f->sq_max_dim()) * 3.; //- BPG by Liu
+            double maxDim = sqrt(f->sq_max_dim());
+            influRegionRadius = maxDim * (3. + _reconSettings->bpgInfluExtra); //- BPG by Liu hardcoded
 
             f->clear_feature();
-            foundBuilding = true;
-            break;
+
+            geomutils::make_round_poly(global::nullPt, influRegionRadius, _boundingRegion);
+            return maxDim;
         }
     }
-    if (!foundBuilding)
-        throw std::invalid_argument("Point of interest does not belong to any building! "
-                                    "Impossible to determine influence region");
+    throw std::invalid_argument("Point of interest does not belong to any building! "
+                                "Impossible to determine influence region");
+}
 
+void BoundingRegion::calc_influ_region_bpg(const double maxDim) {
+    double influRegionRadius = maxDim * (3. + _reconSettings->bpgInfluExtra); //- BPG by Liu hardcoded
     geomutils::make_round_poly(global::nullPt, influRegionRadius, _boundingRegion);
 }
 
+//-- Boundary related functions
 void BoundingRegion::calc_bnd_bpg(const Polygon_2& influRegionPoly,
                                   const BuildingsPtr& buildings) {
     double angle = std::atan2(Config::get().flowDirection.y(), Config::get().flowDirection.x());
@@ -138,6 +148,16 @@ void BoundingRegion::calc_bnd_bpg(const Polygon_2& influRegionPoly,
     }
     //-- Return the points back to global coordinates
     for (auto& pt : localPoly) _boundingRegion.push_back(geomutils::rotate_pt(pt, angle));
+}
+
+// Check if all points of this region fall within otherRegion
+bool BoundingRegion::is_subset_of(const BoundingRegion& otherRegion) const {
+    auto& poly = otherRegion.get_bounding_region();
+    for (auto& vert : _boundingRegion) {
+        if (!geomutils::point_in_poly(vert, poly))
+            return false;
+    }
+    return true;
 }
 
 Polygon_2& BoundingRegion::get_bounding_region() {