+++ src/Building.h
@@ -29,6 +29,7 @@
 #define CITY4CFD_BUILDING_H
 
 #include "PolyFeature.h"
+#include "BoundingRegion.h"
 #include "Config.h"
 
 class Building : public PolyFeature {
@@ -49,7 +50,11 @@ public:
     void   clip_bottom(const TerrainPtr& terrain);
     void   refine();
     void   translate_footprint(const double h);
-    void   check_feature_scope(const Polygon_2& influRegion);
+    bool   is_part_of(const Polygon_2& otherPoly) const;
+    void   set_reconstruction_rules(const BoundingRegion& reconRegion);
+    void   remove_reconstruction_rules();
+    bool   has_reconstruction_region() const;
+    std::shared_ptr<const Config::ReconRegion> get_reconstruction_settings() const;
     void   set_clip_flag (const bool flag);
     void   mark_as_failed();
     bool   has_failed_to_reconstruct() const;
@@ -70,6 +75,7 @@ protected:
     double               _height;
     bool                 _hasFailed;
     bool                 _clip_bottom = Config::get().clip;
+    std::shared_ptr<const Config::ReconRegion> _reconSettings;
 };
 
 //-- Struct for clipping